refactor(core): extract config discovery engine
